Feat: 프로필 페이지 관심사 선택페이지 UI 수정
Resolves: #75
