+++ src/components/Interest/InterestButton/InterestBtn.tsx
@@ -15,12 +15,13 @@ const InterestBtn = ({
 }: InterestBtnProps) => {
   return (
     <div
-      className={`${bgColor} py-2 px-4 mt-6 rounded-full text-_body1 flex items-center`}
+      className={`${bgColor} py-2 px-4 mt-6 rounded-full text-_body1 flex items-center hover:shadow-md`}
     >
       {/* 아이콘 */}
       <FontAwesomeIcon icon={icon} className={`${textColor} mr-4`} />
       <span className={`${textColor} whitespace-nowrap`}>
-        {checkText.length > 6 ? `${checkText.slice(0, 5)}...` : checkText}
+        {/* {checkText.length > 6 ? `${checkText.slice(0, 5)}...` : checkText} */}
+        {checkText}
       </span>
     </div>
   );
