+++ src/components/Interest/InterestCheck/antdCheckbox.css
@@ -1,0 +1,7 @@
+.checkboxHidden > .ant-checkbox {
+  display: none;
+}
+.checkboxHidden > .ant-checkbox:hover {
+  border-color: blue;
+  background-color: yellow;
+}
