+++ src/components/Interest/InterestHeader/InterestHeader.tsx
@@ -2,9 +2,6 @@ const InterestHeader = () => {
   return (
     <div>
       <h1 className="text-_h1 font-bold">관심사를 선택해주세요</h1>
-      <span className="text-_h3 font-normal text-_neutral-70 ">
-        투두메이트에게 나의 관심사를 보여주세요
-      </span>
     </div>
   );
 };
