+++ src/components/SignUp/SignUpForm/SignUpInput/SignUpInput.tsx
@@ -1,4 +1,4 @@
-import React, { ChangeEvent, FocusEvent, KeyboardEventHandler } from "react";
+import React, { ChangeEvent, FocusEvent } from "react";
 
 type SignUpInputProps = {
   label: string;
