+++ src/index.css
@@ -30,6 +30,7 @@ html {
   padding: 20px 64px;
   background: white;
 }
+
 .btn1 {
   width: 472px;
   height: 68px;
