--- src/pages/Interest/antdCheckbox.css
@@ -1,3 +1,0 @@
-.ant-checkbox {
-  display: none;
-}
