+++ src/pages/Interest/interests.json
@@ -1,40 +1,15 @@
 {
   "interests": [
-    { "id": 1, "name": "시험" },
-    { "id": 2, "name": "스포츠" },
-    { "id": 3, "name": "음악" },
-    { "id": 4, "name": "운동" },
-    { "id": 5, "name": "독서" },
-    { "id": 6, "name": "요가" },
-    { "id": 7, "name": "영화" },
-    { "id": 8, "name": "요리" },
-    { "id": 9, "name": "여행" },
-    { "id": 10, "name": "게임" },
-    { "id": 11, "name": "공부" },
-    { "id": 12, "name": "봉사활동" },
-    { "id": 13, "name": "프로그래밍 기술" },
-    { "id": 14, "name": "미술" },
-    { "id": 15, "name": "사진" },
-    { "id": 16, "name": "산책" },
-    { "id": 17, "name": "수영" },
-    { "id": 18, "name": "등산" },
-    { "id": 19, "name": "자전거타기" },
-    { "id": 20, "name": "바둑" },
-    { "id": 21, "name": "테니스" },
-    { "id": 22, "name": "축구" },
-    { "id": 23, "name": "농구" },
-    { "id": 24, "name": "야구" },
-    { "id": 25, "name": "배구" },
-    { "id": 26, "name": "수집" },
-    { "id": 27, "name": "자격증 취득" },
-    { "id": 28, "name": "서핑" },
-    { "id": 29, "name": "글쓰기" },
-    { "id": 30, "name": "그림그리기" },
-    { "id": 31, "name": "헬스" },
-    { "id": 32, "name": "독후감" },
-    { "id": 33, "name": "쿠킹" },
-    { "id": 34, "name": "베이킹" },
-    { "id": 35, "name": "레저" },
-    { "id": 36, "name": "일기" }
+    { "id": 1, "name": "Javascript" },
+    { "id": 2, "name": "React" },
+    { "id": 3, "name": "프론트엔드" },
+    { "id": 4, "name": "백엔드" },
+    { "id": 5, "name": "알고리즘" },
+    { "id": 6, "name": "Java" },
+    { "id": 7, "name": "Kotlin" },
+    { "id": 8, "name": "Vue" },
+    { "id": 9, "name": "Svelt" },
+    { "id": 10, "name": "기획" },
+    { "id": 11, "name": "Spring" }
   ]
 }
